+++ Sources/KeyboardKit/Locale/LocaleDictionary.swift
@@ -17,6 +17,9 @@ import Foundation
  */
 public struct LocaleDictionary<ItemType> {
     
+    /**
+     Create a dictionary with locale entries.
+     */
     public init(_ dict: [KeyboardLocale: ItemType]) {
         self.dictionary = Dictionary(
             uniqueKeysWithValues: dict.keys.compactMap {
@@ -26,12 +29,29 @@ public struct LocaleDictionary<ItemType> {
         )
     }
     
-    public init(_ dict: [String: ItemType]) {
+    /**
+     Create a dictionary with locale entries.
+     */
+    public init(_ dict: [LocaleIdentifier: ItemType]) {
         self.dictionary = dict
     }
     
-    public let dictionary: [String: ItemType]
+    /**
+     This alias indicates the dictionary key type.
+     */
+    public typealias LocaleIdentifier = String
     
+    /**
+     The locale/valye dicitionary.
+     */
+    public let dictionary: [LocaleIdentifier: ItemType]
+    
+    /**
+     Get a certain value for the provided locale.
+     
+     The lookup will first check using the locale identifier
+     then the language code.
+     */
     public func value(for locale: Locale) -> ItemType? {
         if let item = dictionary[locale.identifier] { return item }
         if let item = dictionary[locale.languageCode ?? ""] { return item }